refactor: introduce renderCompat
