+++ src/components/DynamicWidgets.js
@@ -2,7 +2,7 @@ import { createWidgetMixin } from '../mixins/widget';
 import { EXPERIMENTAL_connectDynamicWidgets } from 'instantsearch.js/es/connectors';
 import { createSuitMixin } from '../mixins/suit';
 import { _objectSpread } from '../util/polyfills';
-import { isVue3, h as _h } from '../util/vue-compat';
+import { isVue3, renderCompat } from '../util/vue-compat';
 
 function getWidgetAttribute(vnode) {
   const props = isVue3
@@ -51,9 +51,8 @@ export default {
       default: undefined,
     },
   },
-  render(createElement) {
+  render: renderCompat(function(h) {
     const components = new Map();
-    const h = isVue3 ? _h : createElement;
     const defaultSlot = isVue3
       ? this.$slots.default && this.$slots.default()
       : this.$slots.default;
@@ -90,7 +89,7 @@ export default {
       { class: [this.suit()] },
       this.state.attributesToRender.map(attribute => components.get(attribute))
     );
-  },
+  }),
   computed: {
     widgetParams() {
       return {
