+++ src/util/vue-compat/index-2.js
@@ -7,6 +7,12 @@ const version = Vue.version;
 
 export { Vue, Vue2, isVue2, isVue3, version };
 
+export function renderCompat(fn) {
+  return function(createElement) {
+    return fn.call(this, createElement);
+  };
+}
+
 // Vue3-only APIs
 export const computed = undefined;
 export const createApp = undefined;
