+++ src/util/vue-compat/index-3.js
@@ -6,3 +6,9 @@ const Vue2 = undefined;
 
 export { createApp, createSSRApp, h, version, nextTick } from 'vue';
 export { Vue, Vue2, isVue2, isVue3 };
+
+export function renderCompat(fn) {
+  return function() {
+    return fn.call(this, Vue.h);
+  };
+}
