+++ test/utils/index.js
@@ -75,10 +75,4 @@ export const createSSRApp = props => {
   }
 };
 
-export function renderCompat(fn) {
-  return function(createElementV2) {
-    return isVue3 ? fn.call(this, h) : fn.call(this, createElementV2);
-  };
-}
-
 export const nextTick = () => (isVue3 ? _nextTick() : Vue2.nextTick());
